Move physicsmodels in folder and split into different files; replace magic string logic
