+++ bin/combinetf2_plot_hists_cov.py
@@ -1,10 +1,8 @@
 #!/usr/bin/env python3
 
 import argparse
-import itertools
 import os
 
-import hist
 import matplotlib.pyplot as plt
 import mplhep as hep
 import numpy as np
@@ -12,7 +10,7 @@ import seaborn as sns
 
 import combinetf2.io_tools
 
-from wums import boostHistHelpers as hh  # isort: skip
+# from wums import boostHistHelpers as hh  # isort: skip
 from wums import logging, output_tools, plot_tools  # isort: skip
 
 
@@ -104,7 +102,7 @@ def parseArgs():
         nargs="+",
         action="append",
         default=[],
-        help='Make plot of physics model prefit and postfit histograms, specifying the model name, followed by the channel name and axis names, e.g. "-m project ch0 eta pt". This argument can be called multiple times',
+        help='Make plot of physics model prefit and postfit histograms, specifying the model name, followed by the cinstance key, e.g. "-m project ch0_eta_pt". This argument can be called multiple times',
     )
     parser.add_argument(
         "--prefit", action="store_true", help="Make prefit plot, else postfit"
@@ -235,95 +233,105 @@ def main():
 
     channel_info = meta["meta_info_input"]["channel_info"]
 
-    models = {p[0]: p[1:] for p in args.physicsModel}
+    hist_cov_key = f"hist_{'prefit' if args.prefit else 'postfit'}_inclusive_cov"
 
-    hist_cov = fitresult[
-        f"hist_{'prefit' if args.prefit else 'postfit'}_inclusive_cov"
-    ].get()
+    if len(args.physicsModel) == 0:
+        models = [["basemodel"]]
+    else:
+        models = args.physicsModel
 
-    if len(channel_info) > 1:
-        # plot full covariance matrix only if it goes across multiple channels
-        plot_matrix(outdir, hist_cov, args, config=config, meta=meta)
+    for margs in models:
+        model = margs[0]
 
-    for channel, info in channel_info.items():
-        axes = info["axes"]
-        start = int(info["start"])
-        stop = int(info["stop"])
+        if model not in fitresult.keys():
+            raise ValueError(f"Model {model} not found in fitresult")
 
-        h_cov = hist_cov[{"x": slice(start, stop), "y": slice(start, stop)}]
+        if len(margs) > 1:
+            instance_key = margs[1]
+            if instance_key not in fitresult[model].keys():
+                raise ValueError(
+                    f"Instance {instance_key} of model {model} not found in fitresult"
+                )
+        else:
+            instance_key = None
 
-        plot_matrix(
-            outdir,
-            h_cov,
-            args,
-            channel,
-            [a.name for a in axes],
-            config=config,
-            meta=meta,
-        )
+        for instance_name, instance in fitresult[model].items():
+            if instance_key is not None and instance_name != instance_key:
+                continue
 
-        selection_axes = [a for a in axes if a.name in args.selectionAxes]
-        if (len(models) and channel in [p[0] for p in models.values()]) or len(
-            selection_axes
-        ) > 0:
-            # reshape into original axes
-            h1d = hist.Hist(*axes)
-            h2d = hh.expand_hist_by_duplicate_axes(
-                h1d,
-                [a.name for a in axes],
-                [f"y_{a.name}" for a in axes],
-                put_trailing=True,
+            h_cov = instance[hist_cov_key].get()
+
+            plot_matrix(
+                outdir,
+                h_cov,
+                args,
+                config=config,
+                meta=meta,
+                suffix=model + "_" + instance_name,
             )
 
-            vals = np.reshape(
-                h_cov.values(), (h_cov.shape[0], *h2d.shape[: len(h2d.shape) // 2])
-            )
-            h2d.values()[...] = np.reshape(vals, h2d.shape)
+    # for channel, info in channel_info.items():
+    #     axes = info["axes"]
+    #     start = int(info["start"])
+    #     stop = int(info["stop"])
 
-        if len(selection_axes) > 0:
-            selection_bins = [
-                np.arange(a.size) for a in axes if a.name in args.selectionAxes
-            ]
-            other_axes = [a.name for a in axes if a not in selection_axes]
+    #     h_cov = hist_cov[{"x": slice(start, stop), "y": slice(start, stop)}]
 
-            for bins in itertools.product(*selection_bins):
-                idxs = {a.name: i for a, i in zip(selection_axes, bins)}
-                idxs.update({f"y_{a.name}": i for a, i in zip(selection_axes, bins)})
-                idxs_centers = {
-                    a.name: (
-                        a.centers[i]
-                        if isinstance(a, (hist.axis.Regular, hist.axis.Variable))
-                        else a.edges[i]
-                    )
-                    for a, i in zip(selection_axes, bins)
-                }
-                h_cov_i = h2d[idxs]
-                suffix = f"{channel}_" + "_".join(
-                    [
-                        f"{a}_{str(i).replace('.','p').replace('-','m')}"
-                        for a, i in idxs_centers.items()
-                    ]
-                )
-                plot_matrix(
-                    outdir, h_cov_i, args, suffix, other_axes, config=config, meta=meta
-                )
+    #     plot_matrix(
+    #         outdir,
+    #         h_cov,
+    #         args,
+    #         channel,
+    #         [a.name for a in axes],
+    #         config=config,
+    #         meta=meta,
+    #     )
 
-    for mname, margs in models.items():
-        hist_cov = fitresult[mname]["_".join(margs[1:])][
-            f"hist_{'prefit' if args.prefit else 'postfit'}_inclusive_cov"
-        ].get()
+    #     selection_axes = [a for a in axes if a.name in args.selectionAxes]
+    #     if (len(models) and channel in [p[0] for p in models.values()]) or len(
+    #         selection_axes
+    #     ) > 0:
+    #         # reshape into original axes
+    #         h1d = hist.Hist(*axes)
+    #         h2d = hh.expand_hist_by_duplicate_axes(
+    #             h1d,
+    #             [a.name for a in axes],
+    #             [f"y_{a.name}" for a in axes],
+    #             put_trailing=True,
+    #         )
 
-        # plot full covariance matrix only if it goes across multiple channels
-        plot_matrix(
-            outdir,
-            hist_cov,
-            args,
-            channel=margs[0],
-            axes=margs[1:],
-            config=config,
-            meta=meta,
-            suffix=mname,
-        )
+    #         vals = np.reshape(
+    #             h_cov.values(), (h_cov.shape[0], *h2d.shape[: len(h2d.shape) // 2])
+    #         )
+    #         h2d.values()[...] = np.reshape(vals, h2d.shape)
+
+    #     if len(selection_axes) > 0:
+    #         selection_bins = [
+    #             np.arange(a.size) for a in axes if a.name in args.selectionAxes
+    #         ]
+    #         other_axes = [a.name for a in axes if a not in selection_axes]
+
+    #         for bins in itertools.product(*selection_bins):
+    #             idxs = {a.name: i for a, i in zip(selection_axes, bins)}
+    #             idxs.update({f"y_{a.name}": i for a, i in zip(selection_axes, bins)})
+    #             idxs_centers = {
+    #                 a.name: (
+    #                     a.centers[i]
+    #                     if isinstance(a, (hist.axis.Regular, hist.axis.Variable))
+    #                     else a.edges[i]
+    #                 )
+    #                 for a, i in zip(selection_axes, bins)
+    #             }
+    #             h_cov_i = h2d[idxs]
+    #             suffix = f"{channel}_" + "_".join(
+    #                 [
+    #                     f"{a}_{str(i).replace('.','p').replace('-','m')}"
+    #                     for a, i in idxs_centers.items()
+    #                 ]
+    #             )
+    #             plot_matrix(
+    #                 outdir, h_cov_i, args, suffix, other_axes, config=config, meta=meta
+    #             )
 
     if output_tools.is_eosuser_path(args.outpath) and args.eoscp:
         output_tools.copy_to_eos(outdir, args.outpath, args.outfolder)
